test(test_timesheet.py): Integration tests for schedule UI colors
